+++ db/migrate/20160617183624_rename_comments_to_replies.rb
@@ -1,0 +1,5 @@
+class RenameCommentsToReplies < ActiveRecord::Migration
+  def change
+    rename_table :comments, :replies
+  end
+end