+++ db/migrate/20160617185916_add_reply_id_to_replies.rb
@@ -1,0 +1,6 @@
+class AddReplyIdToReplies < ActiveRecord::Migration
+  def change
+    add_column :replies, :reply_id, :integer
+    add_index :replies, :reply_id
+  end
+end